+++ lib/openstudio/lib/openstudio_aws_wrapper.rb
@@ -54,7 +54,7 @@ class OpenStudioAwsWrapper
   VALID_OPTIONS = [:proxy, :credentials]
 
   def initialize(options = {}, group_uuid = nil)
-    @group_uuid = group_uuid || (SecureRandom.uuid).gsub('-', '')
+    @group_uuid = group_uuid || (SecureRandom.uuid).delete('-')
 
     @security_groups = []
     @key_pair_name = nil