+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 0.3.2 (Unreleased)
 * Add a stable JSON file that can be used to flag which versions of the server are stable (used by OpenStudio PAT).
 * Add the method `describe_availability_zones` to the root AWS class
 * Prefer use of the access and secret key in the environment variables if defined
+* Support i2 instance ephemeral storage. These instances will take a bit longer to startup because the volumes are not yet created.
 
 Version 0.3.1
 -------------